api-python/requirements.txt
flask==3.0.0
flask-cors==4.0.0
jpype1==1.5.0
mpxj==15.2.0
gunicorn==21.2.0
